Extract a superclass for pools
